backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
openai==1.51.0

gunicorn==23.0.0
psycopg2-binary==2.9.10
